requirements.txt
streamlit>=1.36.0
duckdb>=1.0.0
pandas>=2.2.2
pyarrow>=15.0.2
altair>=5.3.0
python-dateutil>=2.9.0
pytz>=2024.1
six>=1.16.0
requests>=2.31.0
python-dotenv>=1.0.1
urllib3>=2.0.0